# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Imagesystems",
    "description": "Imagesystems \u2014 hotel & real estate image enhancement, generation, and batch pipelines",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/images": {
      "get": {
        "tags": [
          "images"
        ],
        "summary": "List Images",
        "operationId": "list_images_api_images_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ImageDetailResponse"
                  },
                  "title": "Response List Images Api Images Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/scrape/embed-check": {
      "post": {
        "tags": [
          "scrape"
        ],
        "summary": "Embed Check",
        "description": "Check whether a URL can be displayed in an iframe from this app (X-Frame-Options / CSP). Many sites\n(e.g. Expedia) forbid embedding \u2014 the panel would stay blank without this check.",
        "operationId": "embed_check_api_scrape_embed_check_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmbedCheckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmbedCheckResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/scrape/page": {
      "post": {
        "tags": [
          "scrape"
        ],
        "summary": "Scrape Page",
        "description": "Fetch a page once (Zyte or direct HTTP) and list image URLs from that HTML.\nNo additional Zyte calls for /scrape/import-urls \u2014 only this endpoint runs browser extract.",
        "operationId": "scrape_page_api_scrape_page_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScrapePageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScrapePageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/scrape/import-urls": {
      "post": {
        "tags": [
          "scrape"
        ],
        "summary": "Import Urls As Images",
        "description": "Download image URLs (each re-validated for SSRF) and register as Image rows \u2014 same as multipart upload.\nDoes not call Zyte; only HTTP GETs the selected asset URLs (your CDN / origin bandwidth).",
        "operationId": "import_urls_as_images_api_scrape_import_urls_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportUrlsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ImageUploadResponse"
                  },
                  "type": "array",
                  "title": "Response Import Urls As Images Api Scrape Import Urls Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EmbedCheckRequest": {
        "properties": {
          "url": {
            "type": "string",
            "maxLength": 2048,
            "minLength": 4,
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "EmbedCheckRequest"
      },
      "EmbedCheckResponse": {
        "properties": {
          "final_url": {
            "type": "string",
            "title": "Final Url"
          },
          "embed_allowed": {
            "type": "boolean",
            "title": "Embed Allowed"
          },
          "detail": {
            "type": "string",
            "title": "Detail",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "final_url",
          "embed_allowed"
        ],
        "title": "EmbedCheckResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ImageDetailResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "original_filename": {
            "type": "string",
            "title": "Original Filename"
          },
          "width": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width"
          },
          "height": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          },
          "file_size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Size Bytes"
          },
          "mime_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mime Type"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/ImageVersionResponse"
            },
            "type": "array",
            "title": "Versions",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "original_filename",
          "created_at"
        ],
        "title": "ImageDetailResponse"
      },
...
      "ImageUploadResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "original_filename": {
            "type": "string",
            "title": "Original Filename"
          },
          "width": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width"
          },
          "height": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          },
          "file_size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Size Bytes"
          },
          "mime_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mime Type"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/ImageVersionResponse"
            },
            "type": "array",
            "title": "Versions"
          }
        },
        "type": "object",
        "required": [
          "id",
          "original_filename",
          "created_at"
        ],
        "title": "ImageUploadResponse"
      },
      "ImageVersionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "version_type": {
            "type": "string",
            "title": "Version Type"
          },
          "width": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width"
          },
          "height": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          },
          "file_size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Size Bytes"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "scale_factor": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scale Factor"
          },
          "processing_cost_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processing Cost Usd"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "prompt_used": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Used"
          },
          "source_job_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Job Type"
          },
          "generation_params": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generation Params"
          }
        },
        "type": "object",
        "required": [
          "id",
          "version_type",
          "created_at"
        ],
        "title": "ImageVersionResponse"
      },
      "ImportUrlImageHint": {
        "properties": {
          "alt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alt"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          }
        },
        "type": "object",
        "title": "ImportUrlImageHint",
        "description": "Per-URL context from the scrape UI (same order as urls). Helps Groq match Gemini-style names."
      },
      "ImportUrlsRequest": {
        "properties": {
          "urls": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Urls"
          },
          "use_groq_naming": {
            "type": "boolean",
            "title": "Use Groq Naming",
            "description": "When True and user has a saved Groq key, rename each import using Groq (text-only; rules align with export AI naming).",
            "default": false
          },
          "page_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2048
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Url",
            "description": "Final page URL after redirects \u2014 context for listing-style names."
          },
          "image_hints": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ImportUrlImageHint"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Hints",
            "description": "Optional list parallel to urls (alt, source). Must match len(urls) when set."
          }
        },
        "type": "object",
        "required": [
          "urls"
        ],
        "title": "ImportUrlsRequest"
      },
...
      "ScrapePageRequest": {
        "properties": {
          "url": {
            "type": "string",
            "maxLength": 2048,
            "minLength": 4,
            "title": "Url"
          },
          "use_rendered_scrape": {
            "type": "boolean",
            "title": "Use Rendered Scrape",
            "description": "When True: Zyte API (browser HTML) when a key is available, else direct HTTP. (use_firecrawl is a legacy JSON alias for this flag.)",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "ScrapePageRequest"
      },
      "ScrapePageResponse": {
        "properties": {
          "page_url": {
            "type": "string",
            "title": "Page Url"
          },
          "final_url": {
            "type": "string",
            "title": "Final Url"
          },
          "images": {
            "items": {
              "$ref": "#/components/schemas/ScrapedImageOut"
            },
            "type": "array",
            "title": "Images"
          },
          "truncated": {
            "type": "boolean",
            "title": "Truncated",
            "default": false
          },
          "scrape_image_cap": {
            "type": "integer",
            "title": "Scrape Image Cap",
            "description": "Maximum image URLs the server will return for one scan (list may be truncated)."
          }
        },
        "type": "object",
        "required": [
          "page_url",
          "final_url",
          "images",
          "scrape_image_cap"
        ],
        "title": "ScrapePageResponse"
      },
      "ScrapedImageOut": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          },
          "alt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alt"
          },
          "source": {
            "type": "string",
            "title": "Source"
          }
        },
        "type": "object",
        "required": [
          "url",
          "source"
        ],
        "title": "ScrapedImageOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}